requirements.txt
github-copilot-sdk
pyyaml
